Modified the server so that it would read the prompt from a file and serve it up to users when they are both in a room.
Also added the problems directory to make this possible. The problems directory has a directory in it for each problem which in turn contains a prompt and a test file.
Also had to modify the roomctrl and server to show users what room they're in. The solution involves using setTimeout to wait for the room view to load. It is very janky.

Also modified the roomCtrl so that it doesn't automatically display a comment block when it presents a prompt (the prompt test itself should have that)