(refactor) Modified the use and setup of the database so that it works with the production database on sebastian's azure
This includes a dbreset file that can be used to wipe out the remote db. It is also used to regenerate the database locally whenever 'grunt local' is run.

Cut out some fat from schema.sql to make it compliant with the mysql syntax that the azure mysql server accepts. Also removed many useless comments from this file.

Cut out fat from the db.js file and put much of the logic found there in dbreset.js

Got rid of most 'USE' sql statements as they can be replaced by adding a field to the object handed in to the dbConnection that the mysql node moduel uses.

Set up the gruntfile to run the appropriate db reset command before running locally

Minor fix to primetime test to fix it (changed the parameter name of the test)

Added some code to log server errors more neatly in server.js